requirements.txt
pillow
getch
